causal-method/requirements.txt
biopython==1.86
causal-history-extraction==0.1.0
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.3
fonttools==4.61.1
kiwisolver==1.4.9
matplotlib==3.10.8
munkres==1.1.4
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
pip==25.3
pyparsing==3.2.5
PySide6==6.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
shiboken6==6.10.1
six==1.17.0
statsmodels==0.14.6
tornado==6.5.4
tzdata==2025.3
unicodedata2==17.0.0
